[#25065] docdb: Remove per-table placement info from LB PerTableLoadState
Summary:
The load balancer's `PerTableLoadState` has a field for placement info by table id, but there is always only one table in that map (because it is **PerTable**LoadState). This diff refactors the placement to just be a `PlacementInfoPB`. The placement is always valid because it comes from `GetPlacementInfoForTable`, which returns table placement, tablegroup placement, or cluster placement (in decreasing precedence).

I've also removed some of the unnecessary placement arguments.

Fixes #25065
Jira: DB-14196

Test Plan: Existing tests

Reviewers: jhe, zdrudi

Reviewed By: zdrudi

Subscribers: ybase, svc_phabricator

Differential Revision: https://phorge.dev.yugabyte.com/D40172