[PLAT-16169] use a custom thread pool in collectReplicationClusterData
Summary:
runAsync uses a globally shared thread pool the size of Runtime.getRuntime().availableProcessors() - 1. This size can be as low as 4 (e.g. on YBM), and if a request to GET DR config is made for an unhealthy DR pair, the entire pool gets used
for up to 2 minutes, which is the default RPC timeout, blocking all other GET DR requests.
This diff changes to a local per-request thread pool. Also, use a shorter timeout for synchronous paths.

Test Plan: Set up DR, bring a node down on source, GET DR, then get DR for another config with a healthy universe pair while the first request is blocked.

Reviewers: cwang, vbansal, sanketh

Reviewed By: vbansal

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D40227