[BACKPORT 2024.2][#24135] docdb: Do not bump schema version for clone of CQL indexes
Summary:
Original commit: 65fbbc1eda510a49cede7e3c1a3f30a5ad4728ae / D40154
Clone bumps the schema version of tables to ensure that the current schema version is
greater than all schema versions that might exist in the snapshot used for clone. This is
incompatible with CQL indexes which are always written to with a schema_version of 0 (because
schema_version is not set when we create the write op, and renaming / dropping columns in CQL is not
allowed).

This diff prevents bumping the schema version for CQL indexes in ImportSnapshot, and also verifies
that the incoming schema version of CQL indexes is 0.

Fixes #24135
Jira: DB-13029

Test Plan: `./yb_build.sh release --cxx-test tools_yb-admin-snapshot-schedule-test --gtest_filter YbAdminSnapshotScheduleTest.DeleteRowsFromCloneYcql`

Reviewers: zdrudi, slingam

Reviewed By: zdrudi

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D40232