Coerce types on read (#76)
`COPY FROM parquet` is too strict when matching Postgres tupledesc schema to the parquet file schema.
e.g. `INT32` type in the parquet schema cannot be read into a Postgres column with `int64` type.
We can avoid this situation by casting arrow array to the array that is expected by the tupledesc
schema, if the cast is possible. We can make use of `arrow-cast` crate, which is in the same project
with `arrow`. Its public api lets us check if a cast possible between 2 arrow types and perform the cast.

To make sure the cast is possible, we need to do 2 checks:
1. arrow-cast allows the cast from "arrow type at the parquet file" to "arrow type at the schema that is
   generated for tupledesc", (user created custom cast functions at Postgres won't work by arrow-cast)
2. the cast is meaningful at Postgres. We check if there is a  cast  from "Postgres type that corresponds to the arrow type at Parquet file" to "Postgres type at the tupledesc".

With that we can implicitly cast between many types as shown below:
- INT16 => INT32
- UINT32 => INT64
- FLOAT32 => FLOAT64
- LargeUtf8 => UTF8
- LargeBinary => Binary
- Struct, Array, and Map with castable fields, e.g. [UINT16] => [INT64] or struct {'x': UINT16} => struct {'x': INT64}
**NOTE**: Struct fields must always strictly match by name and position.

We can cast below types but with runtime errors e.g. value overflow
- INT64 => INT32
- TIMESTAMPTZ => TIMESTAMP

Closes #67.
Closes #79.