[#22983] CDCSDK: Fix before image bug where null values were being skipped
Summary:
This revision fixes a bug in the before image code flow, because of which null values were getting skipped in CDC records.

In the current implementation, the methods `PopulateBeforeImageForUpdateOp` and `PopulateBeforeImageForDeleteOp` populate the before image tuples in the CDC records. Both these methods had a bug wherein an old tuple was populated only if it had a non-null value. This was causing incomplete (hence incorrect) before image values being sent if the value of one of the columns was null. This diff fixes the bug by appropriately populating a null value in the CDC record if the column value is found to be null.
Jira: DB-11901

Test Plan: ./yb_build.sh --cxx-test integration-tests_cdcsdk_replica_identity-test --gtest_filter CDCSDKReplicaIdentityTest.TestBeforeImageForNullOnNullUpdates

Reviewers: asrinivasan, skumar, vkushwaha, siddharth.shah

Reviewed By: skumar

Subscribers: ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D36322