[#25075] YSQL: ON CONFLICT fails to properly update index
Summary:
We figured out that INSERT .. ON CONFLICT UPDATE failed to properly
update a secondary index. It looks like the problem is because of index
update optimization, which is enabled by default. To mitigate the
problem we disable the optimization by default.

Revert "[#18822] YSQL: Enable optimization to skip redundant update operations in 2024.2 by default"

This reverts commit 92e026f33cd3215d348449d070b1c90f0916d063.
Jira: DB-14205

Test Plan: Jenkins: urgent

Reviewers: kramanathan, smishra

Reviewed By: smishra

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D40300