[#25000] xCluster: Fix xcluster_upgrade-test
Summary:
When AutoFlags are promoted or demoted, the tservers are notified only on the next heartbeats.
The xCluster tets wait for xCluster to receive and process these AutoFlags configs, but do not wait for the source tservers to even receive them.
This fix addresses this issue by reducing the heartbeat time, and forcing a sleep after every AutoFlags change.

Fixes #25000
Jira: DB-14144

Test Plan: xcluster_upgrade-test

Reviewers: jhe, slingam, xCluster

Reviewed By: jhe

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D40189