[#25004] YSQL: Fix the backend type of the physical connection in ysql conn mgr
Summary:
Post pg15 merge into the master, the type/name of backends created for physical connections in transaction pool are coming to be client backend instead of yb-conn-mgr worker connection. This diff fixes this issue.

Fix
Based on value of the guc variable `yb_is_client_ysqlconnmgr`, the type of backend is set to yb-conn-mgr worker connection or client backend. While parsing the startup packet of the transactional backend process, the guc variables coming in the startup packet are not yet processed (they are processed in later stage of initialization of the backend). Therefore, in this diff during parsing of the startup packet itself, `yb_is_client_ysqlconnmgr` guc variable is been set and based on it's correct value the type of backend is assigned.

This diff also fixes one of implementation gap while making the connections sticky when client executes SET stmt to set "session_authorization" and "role" guc variables.
Jira: DB-14148

Test Plan: Jenkins: enable connection manager, all tests

Reviewers: skumar, rbarigidad, vpatibandla, stiwary

Reviewed By: stiwary

Subscribers: svc_phabricator, yql

Differential Revision: https://phorge.dev.yugabyte.com/D40323