[BACKPORT 2.20][#24975] CDC: Dont consider expired and not of interest entries for metrics calculation
Summary:
#####Backport Description
Minor merge conflicts in UpdateMetrics()  and cdcsdk_ysql-test.

#####Original Description
In the existing implementation for metrics, we determine that an entry is not being polled based on the absence of last replication time in state table entry. However with the introduction of consistent snapshot streams this mechanism is no longer valid, since this field is populated for every state table entry at the time of stream creation.

This diff fixes this issue by explicitly verifying if a particular state table entry has expired or become not of interest. If such an entry is found then, we don't consider it for metric calculation. We will explicitly clear the values of metric entity associated with such an entry.

However a thing to note here is, in the scenario where some of the tables are not being polled, we will still consider the tablets of these tables for metrics calculation until the time this table becomes not of interest.
Jira: DB-14114

Original commit: dc701f6128624c23ab05b6fb99d501c25e54651b / D40286

Test Plan:
./yb_build.sh --cxx-test integration-tests_cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestLagMetricWithNotOfInterestTableAndCSStream
./yb_build.sh --cxx-test integration-tests_cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestLagMetricWithExpiredTableAndCSStream

Reviewers: skumar, siddharth.shah, vkushwaha

Reviewed By: vkushwaha

Subscribers: vkushwaha, ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D40289