[PLAT-16061] Update xCluster RBAC permission text
Summary:
This is a simple diff to bring the RBAC permission text for
xCluster (Manage XCluster) in line with the rest of the product
text. In YBA we don't capitalize the x in xCluster, and the
permissions modal for RBAC should not be an exception.

The change is simply changing the text to `Manage xCluster` from
`Manage XCluster`.

Test Plan:
Verfiy text change on local YBA.
{F313924}

Reviewers: hzare, vbansal, cwang

Reviewed By: vbansal

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D40360