[PLAT-15699] Add logic for removing internal responses & request bodies from openapi yaml
Summary:
Prior to this diff we were removing internal paths and schemas, but not responses and
request bodies. This could sometimes lead to errors where an internal path is removed
leaving its corresponding response in the yaml but unused.

This diff adds helper functions to remove responses and request bodies from the openapi
yaml.

Test Plan:
Mark a route as internal along with the corresponding response.
Verify that these internal components are removed form the openapi yaml.

Reviewers: #yba-api-review, sneelakantan

Reviewed By: #yba-api-review, sneelakantan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D40376