[PLAT-16207] YBA installer not exists check not working
Summary: Our check to determine if an install does not exist relies on the presence of a marker file that we no longer write, so we incorrectly think an install doesn't exist when it has been installed using the new state file. This diff also checks if the state can be initialized and if it can be, then we fail the install not exists check.

Test Plan: Try to install multiple times, or do replicated-migrate after an install and ensure the check fails.

Reviewers: dshubin, sanketh

Reviewed By: dshubin

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D40338