[#24883] YSQL: Fix major catalog upgrade with tablegroups, indexes, and permissions
Summary:
There are a number of missing pieces to be able to upgrade an index in a tablegroup that’s owned by a regular user. In particular, though the upgrade normally runs as the superuser, after upstream PG commit a117cebd638dd02e5c2e791c25e43745f233111b, CREATE INDEX runs as the table owner. But during the major catalog upgrade, because ACLs have not yet been restored at the time of CREATE INDEX, this causes an ACL check failure. Work around this by skipping the ACL check while in binary upgrade mode.

In addition, tablegroup ACLs aren’t being established due to an incomplete merge in d61d8401044ab630e6cb74054b6c243996e3a1a7. This manifests as an ordinary user not being able to create a table in a tablegroup after the major catalog upgrade, when the user could do so before the upgrade. Fix/complete the code needed to restore the tablegroup ACLs correctly.

Because this code enables ACL dumping during binary upgrade, it runs into a problem with the pg_stat_statements_reset function where because the function now has parameters, the statements emitted fail. Extension upgrade is not yet implemented, so add a YB_TODO for this to fix it when we fix extension upgrade.
Jira: DB-14004

Test Plan:
Jenkins

On AlmaLinux 8:
  ./yb_build.sh fastdebug --gcc11 --cxx-test pg15_upgrade-test --gtest_filter Pg15UpgradeTest.Tablegroup --with_tests -n 50
  ./yb_build.sh fastdebug --gcc11 --java-test org.yb.pgsql.TestYbBackup#testTablegroupAcls -n 50

The first command line depends on a fix for #25171, or else it has a small failure rate, < 5%. I patched a pre-committed change D40465 to avoid the failures.

Reviewers: telgersma, fizaa

Reviewed By: telgersma

Subscribers: loginov, yguan, yql

Differential Revision: https://phorge.dev.yugabyte.com/D40407