[#25120] CDC: Getchanges response safe time stuck even after reading entire WAL
Summary:
During a CDC QA run, we encountered a data loss scenario where the GetChanges response safe time on a tablet was stuck due to which records with commit time > tablet's stream safe time did not get ship.

In logical replication, in order to ship a record with commit time CT, we need confirmation from all other tablets that they do not have any changes <= CT to maintain consistency.

In the QA run, the requested from_op_id in a getchanges call on a tablet was present in segment 'X'. The majority replicated opid index also pointed to the last index in segment 'X'. But note that 'X+1' was the active segment but it was empty as the tablet had no activity after the log rollover. Currently, we used to mark entire WAL as read when we had read the active segment. But in this case, since the majority replicated opid index was also present in segment 'X', we never read segment 'X+1' which was the active segment. Due to this, we always returned segment X's footer safe time in the GetChanges response as opposed to computed stream safe time (sent when we have read the entire WAL). Due to this, records with greater commit time than this response safe time were not shipped as the Virtual WAL could not get confirmation from this tablet.

Root Cause:
We missed this case where although we had read the entire WAL, our 'marker' for reading entire WAL was not set which led us to send footer's safe time in the response instead of the computed stream safe time.

Additionally, we have removed the check asserting correctness of safe time sent in GetChanges response. This will be added back in a separate diff.
Jira: DB-14270

Test Plan: Jenkins

Reviewers: skumar, sumukh.phalgaonkar

Reviewed By: skumar

Subscribers: ycdcxcluster, ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D40378