[#25171] YSQL: Pg15 upgrade: Fix race with upgrading immediately after yb-master restart
Summary:
When yb-master restarts and we are in the middle of a pg15 upgrade then we mark that upgrade state as failed.
This runs via the `SysCatalogLoaded` which is called after yb-master is opened for user RPC calls.

If we start a new upgrade before the call to `SysCatalogLoaded` is made, it will incorrectly mark the newly started upgrade task as failed.

This change fixes it by moving the check for whether we're in the middle of an upgrade earlier, into `VisitSysCatalog` (`YsqlCatalogConfig::SetConfig`) and using the cached info to mark the upgrade as failed.
Jira: DB-14340

Test Plan: Jenkins

Reviewers: tfoucher

Reviewed By: tfoucher

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D40465