[#25176] DocDB: Improve error msg in GetMasterRegistration
Summary:
When yb-master is loading the it is not allowed to server RPC requests. When `GetMasterRegistration` is called at this time it responds with an error, but logs a confusing message.
`I1203 17:15:21.013115 301404 master_cluster_service.cc:248] Patching role from leader to follower because of: Leader not ready to serve requests (yb/master/scoped_leader_shared_lock.cc:130): Leader not yet ready to serve requests: leader_ready_term_ = -1; cstate.current_term = 3`

Since this is in the yb-master logs it sounds as if the raft role is being changed which is not true. This commit changes the message to:

`I1203 17:15:21.013115 301404 master_cluster_service.cc:248] Leader not ready to serve requests (yb/master/scoped_leader_shared_lock.cc:130): yb-master leader is initializing: leader_ready_term_ = -1; cstate.current_term = 3`

Fixes #25176
Jira: DB-14346

Test Plan: Jenkins

Reviewers: tfoucher

Reviewed By: tfoucher

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D40466