[PLAT-16093] Use replication status error as the table status when possible
Summary:
When managing an xCluster configuration on YBA, users are able to sort and search tables by table status on the
xCluster replicaiton/xCluster DR details page.
For tables in error status, we provide additional information about the error in the `replicationStatusErrors` field.
This field contains an array of errors which apply to the current table.
Prior to this diff, the user has no way of searching for this table by the displayed error status because the table is
associated with `error` status instead of say `Schema mismatch` for example. `Schema mismatch` is the text that
appears to the user as we have added logic at the table status component to replace `error` with the array of
replication status errors for a given table.

To address the issue of sorting and search, we decided to simply represent the table status with the first array
element from replicationStatusError. It is very rare for the replicationStatusError array to have more than one
element and taking the first element will give us something to sort with.

Test Plan:
Verify that the tables in error status no longer show up when searching for `status:error` and that they show up
when searching for `status:<the first replicaiton status error shown here>`.
{F316357}
{F316368}
{F316360}
{F316361}
{F318563}
{F318564}

Reviewers: hzare, kkannan, lsangappa, cwang, vbansal

Reviewed By: hzare

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D40470