[PLAT-12178]Storage config associated with scheduled backups should not allowed to be deleted
Summary:
Throw `BAD_REQUEST` if a backup config associated with any backup schedule is tried to be deleted.
Also deprecated the v1 delete API.

Test Plan:
Manually verified that appropriate error is thrown if a backup config associated with backup schedule is tried to be deleted.
Modified existing UTs for new behavior.

Reviewers: #yba-api-review!, vbansal

Reviewed By: vbansal

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D40506