[#23356] docdb: Don't call `Finished` method of async tasks when reloading sys catalog
Summary:
After winning an election and before loading the sys catalog, the master calls `AbortAndWaitForAllTasksUnlocked`. This method kicks off a call chain to call the `AbortAndReturnPrevState` method of every live task in the master. Unfortunately the `AbortAndReturnPrevState` method calls a method, `Finished`, that the async task framework provides as a hook for subclasses to add extra functionality. We have some tasks which call catalog manager methods that grab the mutex in their `Finished` method. Because we don't have re-entrant mutexes this deadlocks the server.

This diff threads a parameter from the catalog manager to the task abort method to selectively skip calling the `Finished` method when aborting a task under certain circumstances. In all other call sites besides the catalog manager reload, we keep the existing behaviour of calling the `Finished` method.
Jira: DB-12281

Test Plan:
```
./yb_build.sh release --cxx-test-filter-re master_tasks-test --cxx-test master_tasks-test --gtest_filter SingleMasterTasksTest. SkipCallbacksWhenReloadingSysCatalog --test-timeout-sec 60
```

Reviewers: xCluster, hsunder, asrivastava, mhaddad

Reviewed By: asrivastava

Subscribers: slingam, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D37208