[#25202] YSQL: Clean up duplicated tablespace validation logic
Summary:
The tablespace validation logic is currently split and duplicated across multiple files. This diff unifies the logic into `TablespaceParser`, which replaces the vaguely named `PlacementInfoConverter`. It also replaces the custom `PgPlacementInfoPB` message type with the more common `ReplicationInfoPB`, which we convert to later in the call chain anyways.

This diff should make two future changes easier:
1. Adding a check to prevent accidental duplicates in the placement blocks
2. Adding read replica support to tablespaces

Because we unify the parse-time and creation-time checks, this diff could cause tablespace parsing to fail if a cluster has incorrect placement info. The only place where I think this is possible is if the already has extra fields in the replica_placement (i.e. the placement was created incorrectly, and before https://phorge.dev.yugabyte.com/D10277). To mitigate this, if a tablespace fails validation, we now only skip over that tablespace (before this diff, we skipped loading all tablespaces if one failed validation).

**Upgrade/Rollback safety:**
The `PgPlacementInfoPB` message type is replaced by `ReplicationInfoPB` in
`PgValidatePlacementRequestPB`. That RPC is only sent from PG to the local tserver, so the code to
process `PgValidatePlacementRequestPB` should upgrade and rollback at the same time as the calling
code.
Jira: DB-14387

Test Plan: existing tests

Reviewers: kfranz, hsunder

Reviewed By: kfranz

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D40520