[#20372] docdb: Fix nullptr deref when ListMasterServers called on a master
Summary:
Calling the ListMasterServers RPC on a master crashes it today. There is no code in our system that does this,
but it is an easy mistake to make when debugging a cluster and using `yb-ts-cli` to inspect cluster status.

This diff returns an unsupported RPC error instead of crashing the master.
Jira: DB-9365

Test Plan:
```
./yb_build.sh --cxx-test-filter-re master_tablet_service_test  --cxx-test master_tablet_service_test
```

Reviewers: asrivastava, mhaddad, hsunder

Reviewed By: asrivastava, mhaddad, hsunder

Subscribers: hsunder, ybase, slingam

Differential Revision: https://phorge.dev.yugabyte.com/D40337