[#25163] CDC: Use last WAL OP log Hybrid time for computing minimum start time of running transactions for segment footer
Summary:
We introduced changes to persist minimum start time of running txns in the WAL segment footer in this [[ https://phorge.dev.yugabyte.com/D36364 | diff ]]. At the time of rollover, if there are no running transactions, this value from callback refers to the tablet leader safe time. The tablet leader safe time may not necessarily be equal to the last WAL Op's log HT in the current segment in cases where there is a delay in appending of last WAL Op and rollover of this segment. For CDC, we use the footer value as the stream safe time and read all WAL Operations from a segment and stream those whose commit_time <= footer value (`min_start_time_of_running_txns`) of the segment. Since tablet leader safe time can be greater than the last WAL Op's log HT of a closed segment,  we are modifying the logic to persist the minimum of (last WAL OP log HT, start time of running transactions from callback). This will help us establish the invariant that stream safe time of a tablet <= last WAL Op's log HT when we have only read a closed WAL segment and not the entire WAL.

As part of this diff, we are making the following changes:
1. Introduced a new field `last_wal_op_log_ht` in segment footer to track log HT of last appended WAL OP. This gets updated each time we append an entry.
2. Use the minimum of callback value & above mentioned field to compute `min_start_time_running_txns`.
3. Re-added the assertion on response safe time that was introduced in this [[ https://phorge.dev.yugabyte.com/D40325 | diff ]] but later removed since it was not being followed as explained above.

**Upgrade/Rollback safety:**
The newly introduced proto field is guarded under a klocalPersisted auto flag - `store_last_wal_op_log_ht`. This flag is checked before accessing or modifying this field.

Test Plan: Existing CDC tests

Reviewers: skumar, sumukh.phalgaonkar

Reviewed By: skumar, sumukh.phalgaonkar

Subscribers: ycdcxcluster, ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D40459